# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Todo Application REST-API scheme",
        "title": "Golang Todo API",
        "contact": {},
        "version": "1.0"
    },
    "host": "127.0.0.1:5050",
    "basePath": "/api/v1",
    "paths": {
...
        "/users": {
            "get": {
                "description": "Перегляд списка користувачів з опиціональною пагінацією",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Список користувачів",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Розмір сторінки з користувачами",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Зміщення сторінки з користувачами",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/github_com_Mirwinli_golang-todoapp_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_Mirwinli_golang-todoapp_internal_core_transport_http_response.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Створити нового користувача в системі",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Створити користувача",
                "parameters": [
                    {
                        "description": "CreateUser тіло запиту",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_features_users_transport_http.CreateUserRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Успішно створений користувач",
                        "schema": {
                            "$ref": "#/definitions/internal_features_users_transport_http.CreateUserResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/github_com_Mirwinli_golang-todoapp_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_Mirwinli_golang-todoapp_internal_core_transport_http_response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/users/{id}": {
            "get": {
                "description": "Отримання конкретного користувача по його ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Отримання користувача",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID получаємого користувача",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Користувачуспішно знайдений",
                        "schema": {
                            "$ref": "#/definitions/internal_features_users_transport_http.GetUserResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/github_com_Mirwinli_golang-todoapp_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_Mirwinli_golang-todoapp_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_Mirwinli_golang-todoapp_internal_core_transport_http_response.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Видалення існуючого в системі користувача по його ID",
                "tags": [
                    "users"
                ],
                "summary": "Видалення користувача",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID видяляємого користувача",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Успішне видалення користувача"
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/github_com_Mirwinli_golang-todoapp_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_Mirwinli_golang-todoapp_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_Mirwinli_golang-todoapp_internal_core_transport_http_response.ErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "description": "Зміна інформації який вже існує в системі користувача\n### Логіка зміни полів {Three-state-logic}:\n1. **Поле не передано**: `phone_number` игноруєтся,значення в БД не змінюєтся\n2. **Явно передано значеня**: `phone_number`: \"+380974526184\" - встановлює новий номер телефону в БД\n3. **Передано null**: `phone_number`:null - очищує поле в БД\nОбмеження: `full_name` не може бути null",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Зміна користувача",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID змінюючого користувача",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "PatchUser тіло запиту",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_features_users_transport_http.PatchUserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успішно змінений користувач",
                        "schema": {
                            "$ref": "#/definitions/internal_features_users_transport_http.PatchUserResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/github_com_Mirwinli_golang-todoapp_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_Mirwinli_golang-todoapp_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/github_com_Mirwinli_golang-todoapp_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_Mirwinli_golang-todoapp_internal_core_transport_http_response.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "github_com_Mirwinli_golang-todoapp_internal_core_transport_http_response.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "full error text"
                },
                "message": {
                    "type": "string",
                    "example": "short human-readable"
                }
            }
        },
...
        "internal_features_users_transport_http.CreateUserRequest": {
            "type": "object",
            "required": [
                "full_name"
            ],
            "properties": {
                "full_name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 3,
                    "example": "Petro Jackson"
                },
                "phone_number": {
                    "type": "string",
                    "maxLength": 15,
                    "minLength": 10,
                    "example": "+3809998887766"
                }
            }
        },
        "internal_features_users_transport_http.CreateUserResponse": {
            "type": "object",
            "properties": {
                "full_name": {
                    "type": "string",
                    "example": "John Doe Borusivoch"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "phone_number": {
                    "type": "string",
                    "example": "29998887766"
                },
                "version": {
                    "type": "integer",
                    "example": 10
                }
            }
        },
        "internal_features_users_transport_http.GetUserResponse": {
            "type": "object",
            "properties": {
                "full_name": {
                    "type": "string",
                    "example": "John Doe Borusivoch"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "phone_number": {
                    "type": "string",
                    "example": "29998887766"
                },
                "version": {
                    "type": "integer",
                    "example": 10
                }
            }
        },
        "internal_features_users_transport_http.PatchUserRequest": {
            "type": "object",
            "properties": {
                "full_name": {
                    "type": "string",
                    "example": "Maksim Maksimovich"
                },
                "phone_number": {
                    "type": "string",
                    "example": "+3801112223344"
                }
            }
        },
        "internal_features_users_transport_http.PatchUserResponse": {
            "type": "object",
            "properties": {
                "full_name": {
                    "type": "string",
                    "example": "John Doe Borusivoch"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "phone_number": {
                    "type": "string",
                    "example": "29998887766"
                },
                "version": {
                    "type": "integer",
                    "example": 10
                }
            }
        }
    }
}